dense_vector_index/requirements.txt
faiss-cpu
numpy
sklearn
transformers
